requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==21.4b2
blessed==1.20.0
boto3==1.28.50
botocore==1.31.50
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.6
contourpy==1.1.1
cycler==0.11.0
Cython==3.0.2
detectron2 @ https://github.com/facebookresearch/detectron2/archive/v0.6.zip
diffdist==0.1
docker-pycreds==0.4.0
einops==0.3.0
faiss-cpu==1.7.4
filelock==3.12.4
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.20.0
google-auth-oauthlib==1.0.0
gorilla-core==0.2.7.8
gpustat==1.0.0
graphviz==0.8.4
grpcio==1.54.2
huggingface-hub==0.17.1
hydra-core==1.1.1
idna==3.4
imageio==2.31.3
importlib-resources==6.0.1
iniconfig==2.0.0
iopath==0.1.9
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.3
lit==17.0.2
lmdb==1.4.1
loguru==0.7.2
lvis @ https://github.com/lvis-dataset/lvis-api/archive/master.zip
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mxnet-mkl==1.6.0
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.495.46
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.0.2
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.76
pandas==2.1.0
panopticapi @ https://github.com/cocodataset/panopticapi/archive/master.zip
pathspec==0.11.2
pathtools==0.1.2
Pillow==9.5.0
pluggy==1.3.0
pooch==1.7.0
portalocker==2.8.2
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pyDeprecate==0.3.1
pydot==1.4.2
pynvml==11.5.0
pyparsing==3.1.1
pyre-extensions==0.0.23
PySocks==1.7.1
pytest==7.4.2
pytorch-lightning==1.4.2
pytz==2023.3.post1
pyvista==0.42.2
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rsa==4.9
s3transfer==0.6.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
scooby==0.7.4
sentry-sdk==1.31.0
setproctitle==1.3.2
smmap==5.0.1
soupsieve==2.5
stable-diffusion-sdkit==2.1.3
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorrt==8.5.3.1
termcolor==2.3.0
test-tube==0.7.5
threadpoolctl==3.2.0
tifffile==2023.8.30
timm==0.6.11
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
torch==1.13.1+cu116
torch-tensorrt==1.3.0
torchaudio==0.13.1+cu116
torchmetrics==0.6.0
torchvision==0.14.1+cu116
tqdm==4.66.1
transformers==4.26.1
triton==2.0.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.16
vtk==9.2.6
wandb==0.15.10
Werkzeug==2.3.7
xformers==0.0.16
yacs==0.1.8
yapf==0.40.1
yarl==1.9.2
zipp==3.16.2
